ref(server>lobby): Changed lobbyId type `uint32_t` --> `uint16_t`

+++ Server/src/lobby/lobby.hpp
@@ -40,7 +40,7 @@ class Lobby;
 
 namespace lobby {
 
-using Id = uint32_t;
+using Id = uint16_t;
 using Callback = std::function<void(Lobby&)>;
 
 }  // namespace lobby
